+++ src/phonemes/phoneme.ts
@@ -1,0 +1,3 @@
+export const SYMBOL_COLUMN_NAME = "symbol";
+
+export type PhonemeAttributes = {[index: string]: boolean};
