+++ src/phonemes/spoken/consonant.ts
@@ -1,0 +1,11 @@
+/**
+ * 
+ */
+
+export enum ConsonantAttribute {
+
+};
+
+export type Consonant = {
+
+};
