+++ src/phonemes/spoken/spoken-phoneme.ts
@@ -1,5 +1,3 @@
-export type SpokenPhoneme = {[index: string]: boolean};
-
 export type SpokenPhonemeAxis = {
     attributes: string[],
     otherAttributes: string[]
